codegen: Remove explicit dependency on httpclient
Currently the code generator depends explictly on the httpclient because
many methods declare org.apache.http.client.ClientProtocolException as
one of the exceptions that they may throw. But all these methods already
declare that they may throw the more generic java.io.IOException, so the
former is redundant. This patch removes the dependency from the POM and
the exception from the throws declarations.

Change-Id: I8f5fdbb0d82d9606763ef16786438e30ac53a8fc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>