codegen: Make URL, user and password configurable
Currently the main class of the generator hardcodes the URL, the user
name and the password used to connect to the engine in order to retrieve
the XML schema of the entities and the RSDL. This patch makes them
configurable, using the --url, --user and --password command line
options. If these options aren't provided default values are used, so
the class can be used exactly as before.

Change-Id: I95c77cb8dfcdf4d1baeb186376c800d100e923e5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>