codegen: Use XSD instead of classpath scanning
Currently the code generator scans the classpath to find the set of
entity classes generated by the xjc compiler. This means that the
generation has to be done in several steps: first generate the entities,
then compile the result, then generate the decorators and the entry
point, then compile again. These steps are currently managed by the
generator itself, which calls Maven to do the intermediate compilation.
To avoid this issue this patch changes the generator so that it extracts
the set of entity classes from the XSD, using XSOM to parse it.

Change-Id: Ia9c7146b477d421ad296cef376cef7c08d051ad8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>