codegen: Don't require the SDK
Currently the code generator uses some utility classes and some entity
classes from the SDK. This introduces a chicken-egg situation, and it
makes difficult to regenerate the SDK from scratch. This patch copies
the required utility methods from the SDK to the generator, and creates
the entity classes from the XML schema before compiling the generator.
This removes the dependency on the SDK.

Change-Id: I32815467bdd91124daa71c0de377809baa0be718
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>