codegen: Use SDK specific JAXB bindings
Currently the SDK uses the same JAXB global bindings used by the engine,
because used to be identical. But recently (in commit a57a6a) the engine
started to use global bindings tha aren't suitable for the SDK. These
bindings used by tne engine already been moved to a separate file in
commit 8acd13. This means that they won't be present in the XML schema
generated by the engine. But some of these global bindings are needed by
the SDK as well, in particular the setting that enables generation of
"isSet" methods. This patch adds a bindings file specific to the SDK,
that contains only the global bindings needed by the SDK.

Change-Id: I946db48bee9d5b9fc6d33ade2fa26b0891f99aaa
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>