codegen: Overload methods for new headers
Currently the code generator generates a single method containing
parameters for all the available headers. This means that when a new
header is added the old method will be removed and a new one will be
created with one additional parameter. This breaks backwards
compatibility. This patch modifies the code generator so that the old
method will be preserved and a new one will be added with the additional
parameter for the new header.

Change-Id: I1c035983a6f74aec7dd611f2aaa4ba1d1581594e
Bug-Url: https://bugzilla.redhat.com/1073417
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 31557aff1b15d3d543612e2be8e987fe34510d24)