sdk: Add X-Ovirt-Expect and Expect
The HTTP Expect header is rejected by the Apache web server if its
values is anythiong other than 100-continue. To avoid problems the
engine explicitly removes it, and it accepts an alternative
X-Ovirt-Except header, with the same values and semantics. This patch
changes the SDK so that it will automatically add the alternative header
when the user adds the original one.

Change-Id: I7af6ffa51c9e5b04e6f57d749e130424e10b7eaf
Related: https://bugzilla.redhat.com/1093784
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>