sdk: Add method to get entity by string id
Currently the method that allows retrieving an entity by id assumes that
the identifier is a UUID, but this isn't true for some entities. This
patch adds an additional method that allows retrieving entities by id
using a string as the parameter.

Note that this patch is the result of running the modified generator,
whitout any change to the metadata used.

Change-Id: I471eb31452e47897724829b52cd5edbe86b94efa
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>