codegen: Handle plural exceptions
The rules for forming plurals are more complicated than just adding a
"s" to the end of the word. We need to implement them correctly, or have
a table of exceptions. For simplicity this patch introduces a table of
exceptions.

Change-Id: I02d75c47af205f25dfe00c1c98dc0045f5d463f6
Bug-Url: https://bugzilla.redhat.com/1113548
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit c8d3d7484ec08109b49c42c717ace5e187bd23f7)