packaging: Distribution specific RPM spec files
Currently we use a single .spec file template for all the supported
distributions. To support all the distributions it contains many "%if"
statements that change behaviour. This makes it convoluted and
difficult to maintain. To avoid this issue this patch introduces
different spec file templates, one for each supported distribution.

Change-Id: Ifea3318d7ab1a64efa533809c5aacff75bf9067d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>