codegen: Extract metadata from Maven artifacts
Currently we require a live engine to generate the metadata. This
increases the time required to re-generate the SDK, specially when there
are more than one version that needs to be re-generated. This patch
changes the code generator so that it extracts the metadata from the
engine Maven artifacts, thus no live engine is required.

Change-Id: I2112e499f428873ab0ae89dae14740bf9611c652
Bug-Url: https://bugzilla.redhat.com/1185853
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>