codegen: Sort class members alphabetically
Currently the code generator doesn't sort the class members in any
particular order. This means that each time the code is re-generated
things will potentially be in a different order, making it difficult to
visually analyze the changes (with the "diff" tool, for example). To
avoid this issue this patch changes the generator so that class members
will be sorted alphabetically.

Change-Id: I59940162b7fad175bd0bce7585f43cac65acb78a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>