codegen: Use explicit version of XJC
The code generator uses the XJC compiler included in the JDK used to run
it. This means that the version of the compiler may be different and it
may produce different results. This is in particular true for the
versions of the XJC compiler included with Java 7 and Java 8. The
changes from version to version aren't usually problematic, but anyhow
they make it difficult to compare versions of the generated code,
because there is no certainty that the differences are real or just
quirks of the XJC compiler. To reduce the chances of this happening this
patch replaces the call to the external "xjc" command with the call to
the XJC class of a specific version managed by Maven.

Change-Id: I26984881cb68b5dca2a12d5976df80253c65abb1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>