sdk: Code generated with Java 8
A previous change introduced the use of Java 8 for both building and
running the code generator. As a result some of the generated source
files generated by the "xjc" compiler are sligthly different, probably
because the "xjc" compiler iterates hash maps, which have different
order in Java 7 and Java 8. This change is the result of running the
code generator with Java 8, so that in the future these differences
won't get in the middle while regenerating the SDK. The changes don't
affect the semantics of the SDK, just the order of some fields and
methods in the code generated by the "xjc" tool.

Change-Id: Ia3d179cd8a4dad4c200e3e527e9ae99562bcac01
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>