sdk: Always generate the tarball when building RPMs
Currently the tarball is only generated if there were changes in the Makefile
or the .spec template. This means that changes to the SDK itself will not be
included in the resulting RPMs unless they are locally commited, which makes
things very confusing for developers. This patc changes the Makefile so that
the tarball is always regenerated.

Change-Id: I20cfe74995e9aac87a342ddc39577d85873313ba
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>