sdk: Avoid duplicated log appenders
Currently when the debug mode is enabled two appenders are added to the
"org.apache.http" logger, one is added to the root category and another
is added to the "org.apache.http" logger itself. This results in
duplicated log messages. This patch changes the SDK so that only one
appender is added.

Change-Id: Ie825b82a1e491d20104d213522daf6aeaccd2047
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>