sdk: Use a dot to separate release number and date
Currently the top level Makefile of the project generates the RPM
release tag adding the date and short git hash, but it does so without
adding a separator. This means that if the release number is 0.8 and the
date and git hash are 20150506.git012345 the complete release tag will
be 0.820150506.git012345. This makes an incorrect release tag. To solve
that issue this patch modifies the top level Makefile so that it adds a
dot to separate the release number and the date, thus generating a
release tag like 0.8.20150506.git012345.

Change-Id: Ia6c4a6ad8a186d6308b1cf74fbe2e309b0388488
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>