sdk: Always use UTF-8 for encoding text
Currently there are several places in the SDK where XML documents are
converted to text without specifying the character set to use. In this
situation the underlying HTTP library will select a character set that
may not match the character set declared in the header of the XML
document, resulting in incorrectly encoded documents. To solve this
issue this patch changes the SDK so that it always specify explicitly
the UTF-8 character set, when generating the XML documents and when
encoding them.

Change-Id: I560e3b96756ec1ffa1bf0169ba116edc18fa9844
Bug-Url: https://bugzilla.redhat.com/1223186
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>