sdk: Avoid duplicated log appenders
Currently when the debug mode is enabled two appenders are added to the
"org.apache.http" logger, one is added to the root category and another
is added to the "org.apache.http" logger itself. This results in
duplicated log messages. This patch changes the SDK so that only one
appender is added.

This change was done before, in commit a9776fd, but that changed only
the generated code, so the change was lost after regenerating the SDK.
This patch changes it generated code and in the template, so it won't be
lost next time the SDK is regenerated.

Change-Id: Ibf617e61a5d8b0ffefc172454228562078bcf0b5
Related: https://gerrit.ovirt.org/40518
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>