sdk: Support "labels" and "virtualfunctionallowedlabels"
The generator of the SDK derives the name of collection getter methods
from the type of the elements of the collection. For example, the type
of "/hosts/{host:id}/nics/{nic:id}/labels" collection is "Labels", so
the name of the getter in the generated "HostNIC" class is "getLabels".

Recently a new "virtualfunctionallowedlabels" collection has been
introduced, that uses the same "Labels" type. As a result the code
generated for the "HostNIC" class contains two methods named
"getLabels", which generates a compilation error.

The right way to fix this would be to derive the name the collection
getter methods from the names of the collections, not the types. But
this would introduce many backwards compatibility problems, so it is
impossible.

This patch introduces an exception for this specific case.

Change-Id: I35bb52ebc6a3e8f0eb4faa376667813e50981f2f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>