Add automation and packaging scripts
This patch adds the automation scripts and the RPM packaging scripts and
spec files.

In order to be able to package the generated SDK, and in preparation to
be able to publish it in Maven Central, the patch also changes the POM
files of the project, so that the POM of the SDK itself doesn't use the
POM of the project as parent. This also means that the configurations
needed to run the generator have to be moved out of the POM of the SDK,
and into the POM of the generator.

Change-Id: I80d140e9504dc0948dbad1b2c0fa9a1e1b8ac21b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>