Add SSO support
Use SSO service for authentication.

When authenticating the sdk should obtain access_token from SSO and use
Bearer authentication to access API using the access token.

The SSO access token is revoked at the end of the session to terminate
the session on SSO and clean up the engine session.

Change-Id: I93eefb798b7987ec08ba287f08375c739e19a4ea
Bug-Url: https://bugzilla.redhat.com/1285885
Signed-off-by: Ondra Machacek <omachace@redhat.com>