Rename package from "sdk" to "sdk4"
Currently the name of the package of the SDK is "org.ovirt.engine.sdk".
This is identical to the package used in version 3 of the SDK, which may
cause issues for applications that want to use both versions of the SDK
simultaneously. To avoid those potential issues this patch renames the
package to "org.ovirt.engine.sdk4".

Change-Id: I410df66e2b3f29fed6330595866e8aaf7813d430
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>