Update to version 4.0.17 model and 1.0.11 of metamodel
The more relevant changes in this new version of the model is that the
"Status" type has been removed, and replaced by enums or strings. So the
tests and examples need to be updated to use "whatever.status()" instead
of "whatever.status().state()".

The patch also fixes some things that didn't work correctly in some of
the examples.

Change-Id: Iadbfc4b76342ddff7fb62447ceaa0dc8a0db9ce2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>