Generate RPM Maven dependencies
Currently the RPM packaging of the SDK doesn't generate the Maven
dependencies.  This means that other packages that use tools like Xmvn
to build won't work correctly. To avoid that this patch changes the RPM
build process so that it genrates a POM file that can be used in RPM
distributions, and generates the dependencies from that POM file.

Change-Id: I2d273398ede6f7a0f86c0253d71660f8decb366b
Bug-Url: https://bugzilla.redhat.com/1372652
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit c81e06c50b978fcc511151413af66cbc099e251b)