Fix tag names of action parameters
Currently when the SDK generates the XML document to invoke an action
method it assumes that the names of the elements inside the 'action'
match the names of the types of the parameters. This isn't correct,
they should match the names of the parameters, not the names of the
types. For example, the 'Host.iscsi_discover' action method has a parameter
of type 'IscsiDetails' and name 'iscsi', so the tag should be 'iscsi',
not 'iscsi_details'. To solve this issue this patch chages the SDK so
that the implementation of action methods create an instance of the
'Action' type, and render it uusing the 'ActionWriter' class, which
knows how to write it correctly.

Change-Id: I25494bc638e3dc1098b1086bad94901ca03a72c9
Signed-off-by: Ondra Machacek <omachace@redhat.com>