Convert 'all_content' parameter to 'All-Content' header
Older versions of the engine (before 4.1) only accepted the
'all_content' parameter as an HTTP header. To better support those
versions of the engine this patch changes the SDK so that when the
'all_content' parameter is specified its value is copied to the
'All-Content' request header. So a request like this:

  GET /ovirt-engine/api/vms/123?all_content=true

Will be translated into this:

  GET /ovirt-engine/api/vms/123?all_content=true
  All-Content: true

Change-Id: I2b7be557957456cc834e2f22468bb8e8485e54f7
Signed-off-by: Ondra Machacek <omachace@redhat.com>