Clarify the units of `provisioned_size`
This patch changes the example to add a VM disk so that it is clearer
that size of a disk has to be specified in bytes. It also changes the
example to use a more realistic size, 10 GiB, as it currently the size
is 1 MiB, which isn't realistic.

Change-Id: Iae61d5ef6a01f59618ed87d2685b5b48400b96ee
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit f049260bf13219bba13041ad389d0444e11d68eb)