Use DiskAttachment.active explicitly
The 'DiskAttachment.active' attribute isn't mandatory when adding a disk
attachment to a virtual machine, but if it isn't provided the result is
undefined: the disk will be active in some situations, and won't in some
other situations. To avoid that uncertainty it is better to always
include it with the desired value. This patch changes the examples
accordingly.

Change-Id: I2e2882418d592b54bc6de270eaaf22ee0b173548
Related-To: https://bugzilla.redhat.com/1390517
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>