Don't fork the Maven source plugin
Currently the project runs the 'jar' goal of the Maven source plugin.u
A side effect of this is that the plugin explicitly executes the
'generate-sources' phase of the build. As a result the code generation,
which is attached to that 'generate-sources' phase, runs twice. That
doesn't have a big impact, as the generated code is the same, but it
wastes resources. This patch changes the build process so that it uses
the 'jar-no-fork', which doesn't have that negative side effect.

Change-Id: Ia8a7b8859caec444350460c90abfcd461dd443dc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>