Unify automation '*.packages' files
Currentl there is one 'build.packages' file, and then multiple symbolic
links, one for each distribution, except Fedora 25. Because of this the
build in Fedora 25 fails. As all the currently supported distributions
require the same packages, there is no need for multiple files, so this
patch replaces the distribution specific files with distribution
independent files.

Change-Id: I08fc74a41ef09f9896f0636fd9cd4abf66da718c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>