Add support for custom headers and query parameters
Currently there is no way to send to the server custom headers or query
parameters. This patch adds support for that. For example, if the
caller, for whatever the reason, needs to send the 'My: myvalue' header
and the 'your=yourvalue' query parameter to the request to add a virtual
machine, it can be done as follows:

  Connection connection = connection()
       .url("https://engine40.example.com/ovirt-engine/api")
       .user("username@domain")
       .password("password")
       .trustStoreFile("truststore.jks")
       .build();

   # Get the reference to the service that manage the collection
   # of virtual machines and add virtual machine:
   VmsService vmsService = connection.systemService().vmsService();
   vmsService.add()
     .vm(vm(...))
     .header("My", "Header")
     .query("Some", "Value")
     .send()

Change-Id: I0d2ac733f5f0952b5ad384aaddae080c1ff491bc
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 6ce7557f29bbb1105b37ad3ddee32599505a94c9)