Avoid duplicate dependencies
Currently the SDK requires the httpclient and httpcore artifacts, and it
explicitly specifies the versions of both components. The httpclient
itself also requires httpcore, but a different versions. This means that
we are requiring two different versions of the same artifact. This
causes problems for applications that use the Maven enforce plugin to
check that there are no duplicates. To make things simpler for those
applications this patch changes the SDK so that it only requires on
version and it isues the Maven enforce plugin to check it.

Change-Id: I1d94a68a9a44701c3bdccd85b0299ffce2c4c1fd
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>