Fix service inheritance
Currently the code generator doesn't take into account that some
services extend other base services. As a result some services are
missing methods. For example, the 'SystemPermissions' service is missing
the 'add' method that it should inherit from the
'AssignedPermissionsService'.

This patch changes the code generator to address that issue.

Change-Id: Ib725cde173e5febc9030c265d992935c1aeae480
Bug-Url: https://bugzilla.redhat.com/1496954
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>