Add support for method extensions
Currently the SDK doesn't support the methods that are defined in the
specification of the API as extensions of other methods. For example,
there is an `AddFromScratch` method that is an extension of the `Add`
method of the service that manages virtual machines. But the SDK doesn't
have a corresponding `addFromScratch` method. This patch modifies the
code generator to add support for those method extensions.

Change-Id: Ie7c8c165ee31ce334ce49dde4df927a3e71d6b73
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>