Automatically replace bad tokens
Currently when the token used by the SDK is no longer valid, because it
is expired or for any other reason, the following requests just fail.
The only way to recover is to close the connection and open a new one.
That isn't convenient for long running applications. To improve that
this patch modifies the SDK so that it automatically replaces bad
tokens, requesting new ones from the SSO service.

Change-Id: I110dfa63c106179bc8da2a6095eceb610e115668
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1434834
Signed-off-by: Ondra Machacek <omachace@redhat.com>