Generator - handle Enum input-parameters
This patch modifies the code generator to handle enums
as stand-alone input parameters to API requests.

By 'stand alone' what is meant, is that we're not talking
about enum fields within another provided type, for example:

  <disk>
      <format>cow</format>
  </disk>

In this example Disk is the provided type and its 'format'
field is an enum within it of the type DiskFormat. This is a
common scenario which is handled well, and is not the subject
of this patch.

Rather, we are talking about the case where an enum is one
of the 'main' input paramters, meaning it is annotated with

  @In

in ovirt-engine-api-model project.

Until version 4.3.21 of ovirt-engine-api-model, no API request
had an enum type as such a stand-alone input parameter.

with the addition of cluster-upgrade action:

   https://gerrit.ovirt.org/98529

came:

  @In ClusterUpgradeAction upgradeAction();

where ClusterUpgradeAction is an enum. This broke the generator,
and necessitated this patch.

Change-Id: Ia92e6a9139a30c335ad8bd0a54f80dba29865efe
Signed-off-by: Ori_Liel <oliel@redhat.com>
(cherry picked from commit 2ac0f0ee45c812c624db58204cfae850f20f92a0)