Bump to model 4.4.20, metamodel 1.3.4
This patch bumps depenedency to model 4.4.20, metamodel 1.3.4.

Since ovirt-engine-api-model 4.4.18 there is an instance
of a URL parameter which is an enum (ovirt-engine-api-metamodel
has been modified to support this from version 1.3.4).

A minor modification was required to the generator of
the java-sdk for this issue, done as well in the context of this
patch.

Change-Id: Ie9a00eac81cf42361876b29cb58fc63f4397786b
Signed-off-by: Ori_Liel <oliel@redhat.com>